requirements.txt
python-dotenv
flask
validators